backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
opencv-python==4.9.0.80
mtcnn==0.1.1
pillow==10.2.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
scikit-learn==1.4.0
matplotlib==3.8.2
tqdm==4.66.1
pytest==7.4.4
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9